homework_06/web-app/pyproject.toml
[tool.poetry]
name = "web-app"
version = "0.1.0"
description = ""
authors = ["roman.mostafin <roman.mostafin@factory5.ai>"]

[tool.poetry.dependencies]
python = "^3.8"
Flask = "<2"
Flask-SQLAlchemy = "^2.5.1"
Flask-Migrate = "^3.0.1"
psycopg2 = "^2.9.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
